Settingspage gives accurate light data and is able to update the onswitch

+++ HueXamlApp/Connector/Light.cs
@@ -7,9 +7,9 @@ namespace HueXamlApp.Connector
     public class Light
     {
         public string Id { get; set; }
-        public double H { get; set; }
-        public double S { get; set; }
-        public double V { get; set; }
+        public int H { get; set; }
+        public int S { get; set; }
+        public int V { get; set; }
         public bool IsOn { get; set; }
         public SolidColorBrush MyBrush { get; set; }
